[performance] add navigate to redirect paths
