+++ src/app/header/header.component.ts
@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
+
 @Component({
   selector: 'app-header',
   imports: [],
   templateUrl: './header.component.html',
-  styleUrl: './header.component.scss'
+  styleUrl: './header.component.scss',
 })
-export class HeaderComponent {
-
-}
+export class HeaderComponent {}
