+++ src/app/pages/ranking/ranking.component.spec.ts
@@ -2,15 +2,15 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { RankingComponent } from './ranking.component';
 
+
 describe('RankingComponent', () => {
   let component: RankingComponent;
   let fixture: ComponentFixture<RankingComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [RankingComponent]
-    })
-    .compileComponents();
+      imports: [RankingComponent],
+    }).compileComponents();
 
     fixture = TestBed.createComponent(RankingComponent);
     component = fixture.componentInstance;
