+++ src/app/pages/ranking/ranking.component.ts
@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
+
 @Component({
   selector: 'app-ranking',
   imports: [],
   templateUrl: './ranking.component.html',
-  styleUrl: './ranking.component.scss'
+  styleUrl: './ranking.component.scss',
 })
-export class RankingComponent {
-
-}
+export class RankingComponent {}
