--- src/app/pages/utils/manageLocalStorage.util.ts
@@ -1,10 +1,0 @@
-export function getLocalStorageKeyValuye(key: string): String | null {
-	if (!key) console.error("Parâmetro inválido");
-	const keyValue = localStorage.getItem(key);
-	return keyValue;
-}
-
-export function setLocalStorageKeyValue(key: string, value: string): void {
-	if (!key || !value) console.error("Parâmetros inválidos");
-	localStorage.setItem(key, value);
-}
